Movi los tipos de los custom hooks al archivo que contiene los tipos
