+++ src/hooks/useProfileData.ts
@@ -1,20 +1,12 @@
 import { toast } from '@/components/ui/use-toast'
 import { supabase } from '@/supabase/supabase'
-
-type profile = {
-  firstName: string
-  lastName: string
-  credentialId: string | undefined
-  document: string
-  birthdate: string
-  email: string
-}
+import { profile } from '@/types/types'
 
 export const useProfileData = () => {
   return {
     insertProfile: async (credentials: profile) => {
       try {
-        const { data, error } = await supabase
+        const { data } = await supabase
           .from('profile')
           .insert([credentials])
           .select()
